[STYLE] : 🖍️ Shorts ViewHolder에 있던 영상 재생 로직을 Shorts Fragment로 이동
