+++ app/src/main/java/com/example/spartube/shorts/ShortsPageFragment.kt
@@ -2,9 +2,12 @@ package com.example.spartube.shorts
 
 import android.content.Intent
 import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ProgressBar
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.media3.common.util.UnstableApi
@@ -19,9 +22,13 @@ import com.example.spartube.db.MyPageEntity
 import com.example.spartube.shorts.recyclerviewutil.BindingModel
 import com.example.spartube.shorts.recyclerviewutil.CommentSetBindingModel
 import com.example.spartube.shorts.recyclerviewutil.ShortsPageAdapter
+import com.example.spartube.shorts.util.CustomPlayerUiController
 import com.example.spartube.shorts.util.PreferenceUtils
 import com.example.spartube.shorts.util.SnapPagerScrollListener
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.YouTubePlayer
+import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.listeners.AbstractYouTubePlayerListener
+import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.options.IFramePlayerOptions
+import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.utils.YouTubePlayerTracker
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.views.YouTubePlayerView
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -37,9 +44,6 @@ class ShortsPageFragment : Fragment() {
     private val shortPageRecyclerAdapter by lazy {
         ShortsPageAdapter(
             context = requireActivity(),
-            onClickItem = { player ->
-                this.player = player
-            },
             onClickShareView = { model ->
                 shareYoutubeInfo(model)
             },
@@ -48,12 +52,16 @@ class ShortsPageFragment : Fragment() {
             },
             onClickComment = { model ->
                 showCommentsWithBottomSheet(model)
+            },
+            startShortsVideo = { model, youtubePlayerView, progressBar ->
+                startShortsVideo(model, youtubePlayerView, progressBar)
             }
         )
     }
 
     private var nextPageToken: String? = null
     private var player: YouTubePlayer? = null
+    private var duration: Float? = 0f
     private val snapHelper = PagerSnapHelper()
     private val endScrollListener by lazy {
         object : RecyclerView.OnScrollListener() {
@@ -149,7 +157,7 @@ class ShortsPageFragment : Fragment() {
         }
     }
 
-    private fun initRecyclerView() = with(binding) {
+    private fun initRecyclerView() {
         if (isAdded) {
             requireActivity().runOnUiThread {
                 shortPageRecyclerAdapter.addItems(shortsList)
@@ -222,6 +230,62 @@ class ShortsPageFragment : Fragment() {
                 }
             }
         }
+    }
+
+    private fun startShortsVideo(
+        model: BindingModel,
+        youtubePlayerView: YouTubePlayerView,
+        progressBar: ProgressBar
+    ) {
+        val tracker = YouTubePlayerTracker()
+        val customUi = youtubePlayerView.inflateCustomPlayerUi(R.layout.shorts_custom_view)
+        val youtubePlayerListener = object : AbstractYouTubePlayerListener() {
+            override fun onReady(youTubePlayer: YouTubePlayer) {
+                val customPlayerUiController = CustomPlayerUiController(customUi)
+                youTubePlayer.run {
+                    player = this
+                    addListener(tracker)
+                    addListener(customPlayerUiController)
+                    loadVideo("t1Jq8pGZ4gU", 0f)
+//                    loadVideo(model.linkId", 0f)
+                    play()
+                }
+            }
+
+            override fun onCurrentSecond(youTubePlayer: YouTubePlayer, second: Float) {
+                super.onCurrentSecond(youTubePlayer, second)
+                Handler(Looper.getMainLooper()).postDelayed({
+                    duration = tracker.videoDuration
+                    duration?.let {
+                        updateProgressBar(it, progressBar, tracker)
+                    }
+                }, 500)
+            }
+        }
+        val iFramePlayerOptions = IFramePlayerOptions.Builder().apply {
+            controls(0)
+            rel(0)
+            ivLoadPolicy(3)
+            ccLoadPolicy(1)
+            fullscreen(0)
+        }.build()
+        youtubePlayerView.initialize(youtubePlayerListener, false, iFramePlayerOptions)
+    }
+
+    private fun updateProgressBar(
+        duration: Float,
+        progressBar: ProgressBar,
+        player: YouTubePlayerTracker
+    ) {
+        val handler = Handler(Looper.getMainLooper())
+        val runnable = object : Runnable {
+            override fun run() {
+                val currentPosition = player.currentSecond
+                progressBar.progress = ((currentPosition * 100) / duration.toInt()).toInt()
+                handler.postDelayed(this, 500)
+            }
+        }
+        handler.postDelayed(runnable, 0)
     }
 
     private fun filteringOverOneMinutes(duration: String): Boolean = duration.contains("M")
